+++ cryptofeed/exchanges.py
@@ -15,6 +15,7 @@ from cryptofeed.exchange.binance_us import BinanceUS
 from cryptofeed.exchange.bitcoincom import BitcoinCom
 from cryptofeed.exchange.bitfinex import Bitfinex
 from cryptofeed.exchange.bitflyer import Bitflyer
+from cryptofeed.exchange.bithumb import Bithumb
 from cryptofeed.exchange.bitmax import Bitmax
 from cryptofeed.exchange.bitmex import Bitmex
 from cryptofeed.exchange.bitstamp import Bitstamp
@@ -49,6 +50,7 @@ EXCHANGE_MAP = {
     BITCOINCOM: BitcoinCom,
     BITFINEX: Bitfinex,
     BITFLYER: Bitflyer,
+    BITHUMB: Bithumb,
     BITMAX: Bitmax,
     BITMEX: Bitmex,
     BITSTAMP: Bitstamp,